plotter.rc: moved subplot params to rc module
