+++ gwpy/plotter/rc.py
@@ -24,6 +24,7 @@ many configurable parameters both interactive, and in saving to disk.
 """
 
 from matplotlib import (rcParams, rc_params)
+from matplotlib.figure import SubplotParams
 
 from .tex import (USE_TEX, MACROS as TEX_MACROS)
 
@@ -89,3 +90,35 @@ rcParams.update(GWPY_PLOT_PARAMS)
 # fix matplotlib issue #3470
 if rcParams['font.family'] == 'serif':
     rcParams['font.family'] = u'serif'
+
+# -- dynamic subplot positioning ----------------------------------------------
+
+SUBPLOT_WIDTH = {
+    8.: (.15, .88),
+    12.: (.1, .92),
+}
+SUBPLOT_HEIGHT = {
+    4.: (.2, .85),
+    6.: (.13, .9),
+    8.: (.11, .92),
+}
+
+
+def get_subplot_params(figsize):
+    """Return sensible default `SubplotParams` for a figure of the given size
+
+    Returns
+    -------
+    params : `~matplotlib.figure.SubplotParams`
+        formatted set of subplot parameters
+    """
+    w, h, = figsize
+    try:
+        l, r = SUBPLOT_WIDTH[w]
+    except KeyError:
+        l = r = None
+    try:
+        b, t = SUBPLOT_HEIGHT[h]
+    except KeyError:
+        b = t = None
+    return SubplotParams(**{'left': l, 'bottom': b, 'right': r, 'top': t})
